requirements.txt
fastapi[standard]
uvicorn
jinja2
python-multipart
pillow
numpy
scikit-learn